[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,16 +2,16 @@
 
 using namespace std;
 
-void calculate(int A, int B)
+void calculate(long long A, long long B)
 {
-    int result = (A + B) * (A - B);
+    long long result = (A + B) * (A - B);
 
     cout << result << endl;
 }
 
 int main()
 {
-    int A, B;
+    long long A, B;
 
     cin >> A >> B;
